# [SPARK-50687][PYTHON] Optimize the logic to get stack traces for DataFrameQueryContext